preparing project structure
