+++ src/lib.rs
@@ -27,60 +27,24 @@ pub extern crate diffsl8_0 as diffsl;
 #[cfg(feature = "diffsl-llvm9")]
 pub extern crate diffsl9_0 as diffsl;
 
-pub trait Scalar:
-    nalgebra::Scalar
-    + From<f64>
-    + Display
-    + SimdRealField
-    + ComplexField
-    + Copy
-    + ClosedSub
-    + From<f64>
-    + ClosedMul
-    + ClosedDiv
-    + ClosedAdd
-    + Signed
-    + PartialOrd
-    + Pow<Self, Output = Self>
-    + Pow<i32, Output = Self>
-{
-    const EPSILON: Self;
-    const INFINITY: Self;
-    const NAN: Self;
-    fn is_nan(self) -> bool;
-}
-
-type IndexType = usize;
-
-impl Scalar for f64 {
-    const EPSILON: Self = f64::EPSILON;
-    const INFINITY: Self = f64::INFINITY;
-    const NAN: Self = f64::NAN;
-    fn is_nan(self) -> bool {
-        self.is_nan()
-    }
-}
-
 pub mod jacobian;
 pub mod linear_solver;
 pub mod matrix;
 pub mod nonlinear_solver;
 pub mod ode_solver;
 pub mod op;
+pub mod scalar;
 pub mod solver;
 pub mod vector;
 
-use std::fmt::Display;
-
 use linear_solver::{lu::LU, LinearSolver};
 use matrix::{DenseMatrix, Matrix, MatrixViewMut};
-use nalgebra::{ClosedAdd, ClosedDiv, ClosedMul, ClosedSub, ComplexField, SimdRealField};
 use nonlinear_solver::{newton::NewtonNonlinearSolver, NonLinearSolver};
-use num_traits::{Pow, Signed};
 pub use ode_solver::{
     bdf::Bdf, equations::OdeEquations, OdeSolverMethod, OdeSolverProblem, OdeSolverState,
 };
-use op::NonLinearOp;
+use op::{LinearOp, NonLinearOp};
+use scalar::{IndexType, Scalar};
 use solver::SolverProblem;
 use vector::{Vector, VectorIndex, VectorRef, VectorView, VectorViewMut};
