+++ src/vector/serial.rs
@@ -15,14 +15,17 @@ impl VectorIndex for DVector<IndexType> {
 
 impl<T: Scalar> VectorCommon for DVector<T> {
     type T = T;
+    type O = T;
 }
 
 impl<'a, T: Scalar> VectorCommon for DVectorView<'a, T> {
     type T = T;
+    type O = T;
 }
 
 impl<'a, T: Scalar> VectorCommon for DVectorViewMut<'a, T> {
     type T = T;
+    type O = T;
 }
 
 impl<'a, T: Scalar> VectorView<'a> for DVectorView<'a, T> {
