+++ src/vector/vector_faer.rs
@@ -76,12 +76,15 @@
 
 // impl VectorCommon for faer::Col<f64> {
 //     type T = f64;
+//     type O = faer::Scale<f64>;
 // }
 // impl<'a> VectorCommon for faer::ColRef<'a, f64> {
 //     type T = f64;
+//     type O = faer::Scale<f64>;
 // }
 // impl<'a> VectorCommon for faer::ColMut<'a, f64> {
 //     type T = f64;
+//     type O = faer::Scale<f64>;
 // }
 
 // impl<'a> VectorView<'a> for faer::ColRef<'a, f64> {
